hooked up AdminOrderHistory component
