+++ client/routes.js
@@ -16,8 +16,8 @@ import Account from './components/Account'
 import PaymentCheckout from './components/cart/PaymentCheckout'
 import OrderToDB from './components/cart/OrderToDB';
 import Help from './components/Help'
-import OrderHistory from '../client/components/OrderHist'
-import AdminOrderHistory from '../client/components/AdminOrderHist'
+import OrderHistory from '../client/components/OrderHistory'
+import AdminOrderHistory from '../client/components/AdminOrderHistory'
 
 /**
  * COMPONENT
@@ -28,11 +28,12 @@ class Routes extends Component {
   }
 
   render() {
-    const {isLoggedIn} = this.props // <= Look at this!!!!!
+    const {isLoggedIn} = this.props
 
     return (
       <Switch>
         {/* Routes placed here are available to all visitors */}
+        <Route exact path="/admin/orders" component={AdminOrderHistory} />
         <Route exact path='/admin/products/add' component={CreateProduct} />
         <Route path='/admin/products/:productId' component={EditProduct} />
         <Route exact path='/products/:productId' component={SingleProduct} />
@@ -41,12 +42,11 @@ class Routes extends Component {
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
         <Route path='/cart/checkout/payment' component={PaymentCheckout} />
-        <Route exact path="/confirmation/" component={OrderToDB} />
-        <Route exact path="/cart/checkout/" render={routeProps => <CheckoutForm {...routeProps} user={this.props.user} />} />
-        <Route path="/our-story/" component={OurStory} />
-        <Route path="/contact-us/" component={Help} />
-        <Route exact path="/cart/" component={CartPage} />
-        <Route exact path="/admin/orders/" component={AdminOrderHistory} />
+        <Route exact path="/confirmation" component={OrderToDB} />
+        <Route exact path="/cart/checkout" render={routeProps => <CheckoutForm {...routeProps} user={this.props.user} />} />
+        <Route path="/our-story" component={OurStory} />
+        <Route path="/contact-us" component={Help} />
+        <Route exact path="/cart" component={CartPage} />
 
         {isLoggedIn && (
           <Switch>
